Replace "must have paid any fee or other money payable..." with "must not have any overdue debts".
This means that having an outstanding invoice does not prevent voting, so long as the invoice is not overdue.
(Q: does this present a corner case when a new member's application has been accepted, but they have not yet paid their first invoice for membership fees?)

I considered using the legal term "in arrears", but settled for the Plain English wording "overdue debts".